VB-1748 - move SessionScheduleDto to dto.visit.scheduler package.
